requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cloudpickle==0.5.3
coincurve==9.0.0
cryptography==2.2.2
cv2wrap==1.0
cycler==0.10.0
dask==0.18.1
decorator==4.3.0
elasticsearch==6.2.0
ffprobe==0.5
Flask==1.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
networkx==2.1
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.1.0
psutil==5.4.7
pycparser==2.18
pycurl==7.43.0.2
pydub==0.22.1
pymongo==3.7.2
pyparsing==2.2.0
PyQt5==5.10.1
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
redis==2.10.6
reedsolo==0.3
Resound==1.2
scikit-image==0.14.0
scipy==1.1.0
sip==4.19.8
six==1.11.0
toolz==0.9.0
urllib3==1.22
Werkzeug==0.14.1

git+ssh://git@gitlab.int.vncsp.com/fromil/elastLog.git#egg=elastLog
git+ssh://git@gitlab.int.vncsp.com/fromil/vncCrypto.git#egg=vncCrypto
git+ssh://git@gitlab.int.vncsp.com/fromil/netTools.git#egg=netTools
git+ssh://git@gitlab.int.vncsp.com/fromil/jsonMaker.git#egg=jsonMaker
